moved global search completer to a new search_completer module and put backwards compatible code in.
