+++ python/search_completer/global_search_completer.py
