+++ python/search_completer/resources/build_resources.sh
