+++ python/search_completer/resources/keyboard.png
