+++ python/search_completer/resources/loading.png
