+++ python/search_completer/resources/modes.graffle
