+++ python/search_completer/resources/no_match.png
