+++ python/search_completer/resources/no_thumbnail.png
