+++ python/search_completer/resources/rect_512x400.png
