+++ python/search_completer/resources/resources.qrc
