+++ python/search_completer/resources/search_result_widget.ui
