+++ python/search_completer/search_result_delegate.py
