+++ python/search_completer/search_result_widget.py
