+++ python/search_completer/ui/resources_rc.py
