+++ python/search_completer/ui/search_result_widget.py
