+++ python/search_completer/utils.py
